#aplite improvements
pre-load the error window so it can be shown in low memory conditions

additional checks for low/out of memory conditions

move next/temp arrivals to appdata to enable more aggressive cleanup

trim down error_window to not use an icon for aplite

+++ src/arrivals.c
@@ -71,7 +71,7 @@ Arrival ArrivalConstructor(const char* trip_id,
                            const char* predicted_arrival, 
                            const char* delta_string, 
                            const int32_t delta, 
-                           const uint32_t bus_index,
+                           const uint8_t bus_index,
                            const char arrival_code) {
                               
   Arrival arrival;
